Movement OK (fixed issues with position calculation and edges)

+++ Asteroids/SpacePoint.h
@@ -4,6 +4,6 @@ struct SpacePoint
 	SpacePoint();
 
 	void add(SpacePoint spaceDistance);
-	int X;
-	int Y;
+	double X;
+	double Y;
 };
